requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
google-generativeai==0.3.2
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
scikit-learn==1.3.1
numpy>=1.24.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
